+++ server/category.hpp
@@ -3,6 +3,7 @@
 
 #include <string>
 #include <vector>
+#include <tuple>
 
 #include "database.hpp"
 #include "user.hpp"
@@ -82,119 +83,21 @@ public:
 
     static Category find(Database &database, uint64_t id)
     {
-        MYSQL_STMT* stmt = database.statement(
-                    "SELECT `id`, `create`, `modify`, `userid`, `name`, `description` "
-                    " FROM `categories` WHERE id = ?");
-        if (stmt == NULL)
-            throw std::logic_error("Can't create statement");
-        
-        unsigned long size[6];
-        uint64_t qid = 0;
-        MYSQL_TIME create;
-        MYSQL_TIME modify;
-        uint64_t quserid;
-        char qname[256+1];
-        char qdesc[256+1];
-        
-        MYSQL_BIND result[6] = {
-            mbind(qid, size[0]),
-            mbind(create, size[1]),
-            mbind(modify, size[2]),
-            mbind(quserid, size[3]),
-            mbind_fixed((char*)qname, size[4], sizeof(qname)),
-            mbind_fixed((char*)qdesc, size[5], sizeof(qdesc)),
-        };
-        static_assert(sizeof(size)/sizeof(size[0]) ==
-                sizeof(result)/sizeof(result[0]),
-                "result and size array should have the same length");
-
-        unsigned long psize[1];
-            MYSQL_BIND param[1] = {
-                mbind(id, psize[0])
-            };
-        
-        if (mysql_stmt_bind_param(stmt, param))
-                throw std::logic_error(
-                        std::string("Can't bind param in statement: ")
-                        +mysql_stmt_error(stmt));
-        if (mysql_stmt_bind_result(stmt, result))
-            throw std::logic_error(
-                    std::string("Can't bind result in statement: ")
-                    +mysql_stmt_error(stmt));
-        if (mysql_stmt_execute(stmt))
-            throw std::logic_error(
-                    std::string("Can't execute statement: ")
-                    +mysql_stmt_error(stmt));
-        // Fetch data
-        auto err = mysql_stmt_fetch(stmt);
-        log_assert(err == 0 || err == MYSQL_NO_DATA);
-        if (err == 0)
-        {
-            log_assert_equal(id, qid);
-            // Name is uniq so we should't get more than one result
-            log_assert_equal(mysql_stmt_fetch(stmt), MYSQL_NO_DATA);
-
-            // Make sure all the results fetched
-            while (mysql_stmt_fetch(stmt) == 0);
-
-            return Category(database, qid,
-                    mysql_to_time(create), mysql_to_time(modify),
-                    quserid, std::string(qname), std::string(qdesc));
-        }
-        return Category();
+        typedef std::tuple<uint64_t> Params;
+        Params params = std::make_tuple(id);
+        return query_uniqe(database, params,
+                "SELECT `id`, `create`, `modify`, `userid`, `name`, `description` "
+                " FROM `categories` WHERE id = ?");
     }
 
     static void find_all(Database &database,
             User const &user, std::vector<Category> &categories)
     {
-        MYSQL_STMT* stmt = database.statement(
-                    "SELECT `id`, `create`, `modify`, `name`, `description` "
-                    " FROM `categories` WHERE userid = ?");
-        if (stmt == NULL)
-            throw std::logic_error("Can't create statement");
-        unsigned long size[5];
-        uint64_t qid = 0;
-        MYSQL_TIME create;
-        MYSQL_TIME modify;
-        char qname[64+1];
-        char qdesc[256+1];
-        
-        MYSQL_BIND result[5] = {
-            mbind(qid, size[0]),
-            mbind(create, size[1]),
-            mbind(modify, size[2]),
-            mbind_fixed((char*)qname, size[3], sizeof(qname)),
-            mbind_fixed((char*)qdesc, size[4], sizeof(qdesc)),
-        };
-        static_assert(sizeof(size)/sizeof(size[0]) ==
-                sizeof(result)/sizeof(result[0]),
-                "result and size array should have the same length");
-
-        uint64_t userid = user.id();
-        unsigned long psize[1];
-        MYSQL_BIND param[1] = {
-            mbind(userid, psize[0])
-        };
-        
-        if (mysql_stmt_bind_param(stmt, param))
-                throw std::logic_error(
-                        std::string("Can't bind param in statement: ")
-                        +mysql_stmt_error(stmt));
-        if (mysql_stmt_bind_result(stmt, result))
-            throw std::logic_error(
-                    std::string("Can't bind param in statement: ")
-                    +mysql_stmt_error(stmt));
-        if (mysql_stmt_execute(stmt))
-            throw std::logic_error(
-                    std::string("Can't execute statement: ")
-                    +mysql_stmt_error(stmt));
-        while (mysql_stmt_fetch (stmt) == 0)
-        {
-            categories.push_back(Category(database, qid,
-                        mysql_to_time(create), mysql_to_time(modify),
-                        user.id(),
-                        std::string(qname), std::string(qdesc)));
-        }
+        typedef std::tuple<uint64_t> Params;
+        Params params = std::make_tuple(user.id());
+        query(database, params, categories,
+                "SELECT `id`, `create`, `modify`, `userid`, `name`, `description` "
+                " FROM `categories` WHERE userid = ?");
     }
 
 private:
@@ -207,6 +110,45 @@ private:
         _id(id), _create(create), _modify(modify),
         _userid(userid), _name(name), _description(description)
     {}
+
+    template <typename Params>
+    static Category query_uniqe(Database &database, Params const &params,
+            std::string const &query)
+    {
+        typedef std::tuple<uint64_t, MYSQL_TIME, MYSQL_TIME,
+                uint64_t, FixedString<128>, FixedString<256>> Results;
+        Category category;
+        database.query<Params, Results>(params, query,
+                [&database, &category] (Results &res) {
+                    log_assert(!category.valid());
+                    category = Category(database,
+                        std::get<0>(res),
+                        mysql_to_time(std::get<1>(res)),
+                        mysql_to_time(std::get<2>(res)),
+                        std::get<3>(res),
+                        std::get<4>(res).to_string(),
+                        std::get<5>(res).to_string());
+                });
+        return category;
+    }
+
+    template <typename Params>
+    static void query(Database &database, Params const &params,
+            std::vector<Category> &categories, std::string const &query)
+    {
+        typedef std::tuple<uint64_t, MYSQL_TIME, MYSQL_TIME,
+                uint64_t, FixedString<128>, FixedString<256>> Results;
+        database.query<Params, Results>(params, query,
+                [&database, &categories] (Results &res) {
+                    categories.push_back(Category(database,
+                        std::get<0>(res),
+                        mysql_to_time(std::get<1>(res)),
+                        mysql_to_time(std::get<2>(res)),
+                        std::get<3>(res),
+                        std::get<4>(res).to_string(),
+                        std::get<5>(res).to_string()));
+                });
+    }
 
     void lazy_load()
     {