+++ server/database.hpp
@@ -6,6 +6,7 @@
 #include <stdexcept>
 #include <cstring>
 #include <time.h>
+#include <functional>
 #include <mysql/mysql.h>
 
 #include "logger.hpp"
@@ -22,83 +23,24 @@
     }
 };*/
 
-class Database
+template <size_t size>
+class FixedString
 {
+    char _c_str[size+1];
+
 public:
-    Database(std::string const &server, std::string const &user,
-            std::string const &pass, std::string const &database)
+    FixedString() { _c_str[0] = 0; }
+    FixedString(std::string const &str)
     {
-        // connect to database
-        _mysql = mysql_init(NULL);
-        if (_mysql == NULL) 
-        {
-            _error_message = "Can't initialize mysql: ";
-            _error_message += mysql_error(_mysql);
-        } else
-        if (mysql_real_connect(_mysql, server.c_str(), user.c_str(),
-                    pass.c_str(), database.c_str(),
-                    0, NULL, 0) == NULL) 
-        {
-            _error_message = "Can't connect to database: ";
-            _error_message += mysql_error(_mysql);
-        } else _connected = true;
+        memcpy(_c_str, str.c_str(), str.length()+1);
     }
-
-    ~Database()
+    char const* c_str() const { return _c_str; }
+    char* c_str() { return _c_str; }
+    size_t length() const { return strlen(_c_str); }
+    std::string to_string()
     {
-        // Make sure all the prepared statements closed.
-        // Ignore errors, we need to close them anyway, and it is
-        // also bad practice to throw exception in dtor.
-        for (auto &it : statements) mysql_stmt_close(it.second);
-        // close database connection
-        mysql_close(_mysql);
+        return std::string(_c_str);
     }
-
-    Database(Database const&) = delete;
-    Database& operator=(Database const&) = delete;
-
-    bool connected() { return _connected; }
-
-    std::string error_message() { return _error_message; }
-
-    bool table_exists(std::string const &name)
-    {
-        return mysql_query(_mysql, ("SELECT count(*) from `"+name+"`").c_str()) == 0;
-    }
-
-    MYSQL_STMT* statement(std::string const &query)
-    {
-        auto it = statements.find(query);
-        if (it == statements.end())
-        {
-            MYSQL_STMT *stmt = mysql_stmt_init(_mysql);
-            if (!stmt) throw std::logic_error("Prepare statement init falied!");
-            if (mysql_stmt_prepare(stmt, query.c_str(), query.length()))
-            {
-                throw std::logic_error(std::string("Prepare statement failed: ")+
-                        +mysql_stmt_error(stmt));
-            }
-            statements[query] = stmt;
-            return stmt;
-        }
-        return it->second;
-    }
-
-    /*void query(std::string const &query, MYSQL_BIND *result)
-    {
-    }*/
-
-    uint64_t last_insert_id()
-    {
-        return mysql_insert_id(_mysql);
-    }
-
-private:
-
-    std::map<std::string, MYSQL_STMT*> statements;
-    bool _connected;
-    MYSQL *_mysql;
-    std::string _error_message;
 };
 
 inline time_t mysql_to_time(MYSQL_TIME &dt)
@@ -164,6 +106,21 @@ inline MYSQL_BIND mbind(std::string const &str, unsigned long &size)
     return bind;
 }
 
+template <size_t Size>
+MYSQL_BIND mbind(FixedString<Size> const &str, unsigned long &size)
+{
+    size = str.length();
+    LOG_DEBUG("Lemngth: %d", size);
+    MYSQL_BIND bind;
+    memset(&bind, 0, sizeof(bind));
+    bind.buffer_type = MYSQL_TYPE_STRING;
+    bind.buffer = (void*)str.c_str();
+    bind.buffer_length = Size;
+    bind.length = &size;
+    bind.is_null = 0;
+    return bind;
+}
+
 inline MYSQL_BIND mbind(uint64_t const &num, unsigned long &size)
 {
     size = sizeof(num);
@@ -190,5 +147,166 @@ inline MYSQL_BIND mbind(MYSQL_TIME const &datetime, unsigned long &size)
     return bind;
 }
 
+class Database
+{
+    template <class BindTuple, size_t N>
+    struct SetupBind
+    {
+        static void setup(BindTuple const &tuple, size_t *size, MYSQL_BIND *bind)
+        {
+            SetupBind<BindTuple, N-1>::setup(tuple, size-1, bind-1);
+            //LOG_DEBUG("bind param to %X", (uint64_t)bind);
+            *bind = mbind(std::get<N-1>(tuple), *size);
+        }
+    };
+
+    template <class BindTuple>
+    struct SetupBind<BindTuple, 1>
+    {
+        static void setup(BindTuple const &tuple, size_t *size, MYSQL_BIND *bind)
+        {
+            //LOG_DEBUG("bind last param to %X", (uint64_t)bind);
+            *bind = mbind(std::get<0>(tuple), *size);
+        }
+    };
+
+    template <class BindTuple>
+    struct SetupBind<BindTuple, 0>
+    {
+        static void setup(BindTuple const &tuple, size_t *size, MYSQL_BIND *bind)
+        {
+            (void)tuple; (void)size; (void)bind;
+        }
+    };
+
+public:
+    Database(std::string const &server, std::string const &user,
+            std::string const &pass, std::string const &database)
+    {
+        // connect to database
+        _mysql = mysql_init(NULL);
+        if (_mysql == NULL) 
+        {
+            _error_message = "Can't initialize mysql: ";
+            _error_message += mysql_error(_mysql);
+        } else
+        if (mysql_real_connect(_mysql, server.c_str(), user.c_str(),
+                    pass.c_str(), database.c_str(),
+                    0, NULL, 0) == NULL) 
+        {
+            _error_message = "Can't connect to database: ";
+            _error_message += mysql_error(_mysql);
+        } else _connected = true;
+    }
+
+    ~Database()
+    {
+        // Make sure all the prepared statements closed.
+        // Ignore errors, we need to close them anyway, and it is
+        // also bad practice to throw exception in dtor.
+        for (auto &it : statements) mysql_stmt_close(it.second);
+        // close database connection
+        mysql_close(_mysql);
+    }
+
+    Database(Database const&) = delete;
+    Database& operator=(Database const&) = delete;
+
+    bool connected() { return _connected; }
+
+    std::string error_message() { return _error_message; }
+
+    bool table_exists(std::string const &name)
+    {
+        return mysql_query(_mysql, ("SELECT count(*) from `"+name+"`").c_str()) == 0;
+    }
+
+    MYSQL_STMT* statement(std::string const &query)
+    {
+        auto it = statements.find(query);
+        if (it == statements.end())
+        {
+            MYSQL_STMT *stmt = mysql_stmt_init(_mysql);
+            if (!stmt) throw std::logic_error("Prepare statement init falied!");
+            if (mysql_stmt_prepare(stmt, query.c_str(), query.length()))
+            {
+                throw std::logic_error(std::string("Prepare statement failed: ")+
+                        +mysql_stmt_error(stmt));
+            }
+            statements[query] = stmt;
+            return stmt;
+        }
+        return it->second;
+    }
+
+    template <class Params, class Results,
+             typename ResultCallback = std::function< void (Results &res)>>
+    void query(Params const &param_values, std::string const &query, ResultCallback cb)
+    {
+        // Get statement
+        MYSQL_STMT *stmt = statement(query);
+        // Setup query arguments
+        //Params param_values;
+        size_t param_size[std::tuple_size<Params>::value];
+        MYSQL_BIND params[std::tuple_size<Params>::value];
+        SetupBind<Params, std::tuple_size<Params>::value>::setup(
+                param_values,
+                param_size+std::tuple_size<Params>::value-1,
+                params+std::tuple_size<Params>::value-1);
+        
+        if (mysql_stmt_bind_param(stmt, params))
+            throw std::logic_error(
+                    std::string("Can't bind param in statement: ")
+                    +mysql_stmt_error(stmt));
+        // Setup result set
+        Results res_values;
+        size_t  res_size[std::tuple_size<Results>::value];
+        MYSQL_BIND results[std::tuple_size<Results>::value];
+        SetupBind<Results, std::tuple_size<Results>::value>::setup(
+                res_values,
+                res_size+std::tuple_size<Results>::value-1,
+                results+std::tuple_size<Results>::value-1);
+        if (mysql_stmt_bind_result(stmt, results))
+            throw std::logic_error(
+                    std::string("Can't bind result in statement: ")
+                    +mysql_stmt_error(stmt));
+
+        // Execute query
+        if (mysql_stmt_execute(stmt))
+            throw std::logic_error(
+                    std::string("Can't execute statement: ")
+                    +mysql_stmt_error(stmt));
+        // Fetch data
+        /*for (unsigned i = 0; i<std::tuple_size<Results>::value;++i)
+                LOG_DEBUG("%d size: %d", i, res_size[i]);*/
+        int error;
+        while ((error = mysql_stmt_fetch (stmt)) == 0)
+        {
+            LOG_DEBUG("Got result");
+            /*for (unsigned i = 0; i<std::tuple_size<Results>::value;++i)
+                LOG_DEBUG("%d got size: %d", i, res_size[i]);*/
+            cb(res_values);
+        }
+        if (error != MYSQL_NO_DATA)
+        {
+            LOG_DEBUG("fetch error: %d", error);
+            throw std::logic_error(
+                    std::string("Can't fetch: ")
+                    +mysql_stmt_error(stmt));
+        }
+    }
+
+    uint64_t last_insert_id()
+    {
+        return mysql_insert_id(_mysql);
+    }
+
+private:
+
+    std::map<std::string, MYSQL_STMT*> statements;
+    bool _connected;
+    MYSQL *_mysql;
+    std::string _error_message;
+};
 
 #endif