+++ server/assets/changelog.markdown
@@ -1,5 +1,11 @@
 # FinAlyst changelog #
 
+## 2013. 11. 07. ##
+
+* refactored database query handling
+* fixed assert macro
+* Plan class renamed to PlanItem (no changes in the JSON interface)
+
 ## 2013. 10. 30. ##
 
 * active record objects obtaine id after insert